+++ subgraph/src/ids.ts
@@ -1,5 +1,0 @@
-import {Address} from '@graphprotocol/graph-ts';
-
-export function getDaoId(daoAddress: Address): string {
-  return daoAddress.toHexString();
-}