+++ subgraph/CHANGELOG.md
@@ -11,4 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `getPluginRepoId`, `getPluginSetupId`, `getPluginInstallationId`, `getPluginPreparationId`, `getPluginReleaseId`, `getPluginVersionId`, and `getPluginPermissionId`.
+
 ### Removed